feat: Save Module.symvers to kernel pkg image
Building external kernel modules usually requires a Module.symvers file
corresponding to the target kernel exact build. The file is generated as
one of the last steps in the kernel modules build (modpost), so it is
not feasible to regenerate it in each external module build. Having this
file present in the kernel package makes it easier for third parties to
make Talos system extensions for external modules without having to
recompile the whole kernel in the target configuration.