feat: Kernel config for NVIDIA Tegra drivers
Move some of the Tegra in-kernel drivers from =y to =m to allow using
the NVIDIA out-of-tree drivers as an option. Some of the OOT drivers use
the same module names as the in-kernel drivers, so loading them is
impossible if the in-kernel driver is linked into the kernel itself
already. Compiling the in-kernel drivers as modules allows for better
flexibility.

Signed-off-by: Joakim Nohlgård <joakim@nohlgard.se>